Fix: new/modified content

+++ 26db_drop_table2.py
@@ -14,9 +14,14 @@ else:
     try:
         with con:
             cur = con.cursor()
+            query = "SELECT name FROM sqlite_master WHERE type='table';"
+            allTablesList = [item[0] for item in cur.execute(query).fetchall()]
             for tableName in tableNameList:
-                query = "DROP TABLE {}".format(tableName)
-                cur.execute(query)
-                print("Table dropped:", tableName)
+                if tableName in allTablesList:
+                    query = "DROP TABLE {}".format(tableName)
+                    cur.execute(query)
+                    print("Table dropped:", tableName)
+                else:
+                    print("Table {} doesn't exists".format(tableName))
     except sqlite3.Error as error:
         print("Problem to drop table, {}".format(error))
